Closure_Project/requirements.txt
pandas~=1.2.4
requests
Django~=3.2
bs4~=0.0.1
beautifulsoup4~=4.9.3
urllib3~=1.26.4
jsonpickle~=2.0.0
lxml~=4.6.3
html5lib~=1.1
pytz~=2021.1
sqlparse~=0.4.1
asgiref~=3.3.4
numpy~=1.20.2
setuptools~=56.2.0
djangorestframework~=3.12.4
djangorestframework-jwt==1.11.0
tqdm>=4.61.2
django-cors-headers==3.6.0
drf_yasg==1.20.0
drf-nested-routers~=0.93.3
django-filter~=2.4.0
PyJWT[crypto]~=1.7.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
google-cloud-secret-manager==2.4.0
google-auth==1.30.0
django-storages[google]==1.11.1
django-environ==0.4.5
Collectfast==2.2.0
pytest~=6.2.4
pytest-django~=4.4.0
pytest-dotenv~=0.5.2
